// packages/datadog-api-client-v1/models/FormulaAndFunctionEventQueryDefinitionCompute.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FormulaAndFunctionEventAggregation } from './FormulaAndFunctionEventAggregation';
/**
* Compute options.
*/
export declare class FormulaAndFunctionEventQueryDefinitionCompute {
    'aggregation': FormulaAndFunctionEventAggregation;
    /**
    * A time interval in milliseconds.
    */
    'interval'?: number;
    /**
    * Measurable attribute to compute.
    */
    'metric'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): FormulaAndFunctionEventQueryDefinitionCompute;
    static serialize(data: FormulaAndFunctionEventQueryDefinitionCompute): {
        [key: string]: any;
    };
    constructor();
}