// packages/datadog-api-client-v1/models/FormulaAndFunctionEventQueryGroupBy.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FormulaAndFunctionEventQueryGroupBySort } from './FormulaAndFunctionEventQueryGroupBySort';
/**
* List of objects used to group by.
*/
export declare class FormulaAndFunctionEventQueryGroupBy {
    /**
    * Event facet.
    */
    'facet': string;
    /**
    * Number of groups to return.
    */
    'limit'?: number;
    'sort'?: FormulaAndFunctionEventQueryGroupBySort;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): FormulaAndFunctionEventQueryGroupBy;
    static serialize(data: FormulaAndFunctionEventQueryGroupBy): {
        [key: string]: any;
    };
    constructor();
}